requirements.txt
Flask==1.1.2
beautifulsoup4==4.9.1
gunicorn==20.0.4
jupyter==1.0.0
lxml==4.5.2
pylint==2.5.3
pytest==6.0.1
pytest-cov==2.10.0
PyYAML==5.3.1
requests==2.24.0

